1st release not perfect but working
Clicking into the text onto the lower right 6 different algorithms are being use dto determine the line number of the current cursor position. The text is a lorem ipsum wuth 300000 lines. So the file is already nearly 20MB large. This is about the maximum a NSTextView will handle. At about 400000 lines NSTextView tilts.